requirements.txt
flask-restful
flask
pytesseract
numpy
opencv-python-headless
matplotlib
scikit-image
Pillow
dateparser
gunicorn
